qt: fix compiler warnings/errors on msvc

+++ qt/src/Map.cpp
@@ -5,13 +5,13 @@
 
 using namespace std;
 
-Map::Map(uint8_t layers, uint8_t width, uint8_t height)
+Map::Map(uint8_t layers, uint16_t width, uint16_t height)
   : Map(layers, width, height, nullptr)
 {
 
 }
 
-Map::Map(uint8_t layers, uint8_t width, uint8_t height, uint8_t* data)
+Map::Map(uint8_t layers, uint16_t width, uint16_t height, uint8_t* data)
   : m_layers(layers)
   , m_height(height)
   , m_width(width)
@@ -42,7 +42,7 @@ Map::~Map()
   }
 }
 
-uint8_t Map::getValue(uint8_t layer, uint8_t x, uint8_t y) const
+uint8_t Map::getValue(uint8_t layer, uint16_t x, uint16_t y) const
 {
   if (layer >= m_layers) {
     throw runtime_error("Invalid map layer index");
@@ -55,7 +55,7 @@ uint8_t Map::getValue(uint8_t layer, uint8_t x, uint8_t y) const
   return m_data[y * m_width * m_layers + layer * m_width + x];
 }
 
-void Map::setValue(uint8_t layer, uint8_t x, uint8_t y, uint8_t value)
+void Map::setValue(uint8_t layer, uint16_t x, uint16_t y, uint8_t value)
 {
   if (layer >= m_layers) {
     throw runtime_error("Invalid map layer index");
